+++ lib/utils/config.js
@@ -13,6 +13,7 @@ class Config {
 		apiBase: undefined,
 		token: undefined,
 	};
+	landingPageMdSrc = 'https://github-cdn.now.sh/privatenumber/github-cdn/master/readme.md?token';
 	memcached = undefined;
 	whitelist = undefined;
 
@@ -20,7 +21,14 @@ class Config {
 		const { config } = cosmiconfigSync('githubcdn').search() || {};
 
 		if (config) {
-			Object.assign(this.github, config.github);
+			if (config.github) {
+				Object.assign(this.github, config.github);
+			}
+
+			if (config.landingPageMdSrc) {
+				this.landingPageMdSrc = config.landingPageMdSrc;
+			}
+
 			this.memcached = config.memcached;
 			this.whitelist = config.whitelist;
 		}
